+++ lib/highcharts_stock.dart
@@ -21,6 +21,7 @@ import 'package:flutter/material.dart';
 
 import 'options/highcharts_options.dart';
 import 'utilities/highcharts_callback.dart';
+import 'utilities/highcharts_exporting.dart';
 import 'utilities/highcharts_helpers.dart';
 import 'utilities/highcharts_view.dart';
 
@@ -123,17 +124,18 @@ class HighchartsStock extends StatelessWidget {
     this.debug = kDebugMode,
     this.javaScript,
     this.javaScriptModules = const [
-      'https://code.highcharts.com/stock/highstock.js',
-      'https://code.highcharts.com/stock/highcharts-more.js',
-      'https://code.highcharts.com/stock/highcharts-3d.js',
-      'https://code.highcharts.com/stock/modules/solid-gauge.js',
-      'https://code.highcharts.com/stock/modules/annotations.js',
-      'https://code.highcharts.com/stock/modules/boost.js',
-      'https://code.highcharts.com/stock/modules/broken-axis.js',
-      'https://code.highcharts.com/stock/modules/data.js',
-      'https://code.highcharts.com/stock/modules/exporting.js',
-      'https://code.highcharts.com/stock/modules/offline-exporting.js',
-      'https://code.highcharts.com/stock/modules/accessibility.js',
+      'packages/highcharts_flutter/assets/highcharts/highstock.js',
+      'packages/highcharts_flutter/assets/highcharts/highcharts-more.js',
+      'packages/highcharts_flutter/assets/highcharts/indicators/indicators-all.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/solid-gauge.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/annotations.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/boost.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/broken-axis.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/data.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/exporting.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/offline-exporting.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/export-data.js',
+      'packages/highcharts_flutter/assets/highcharts/modules/accessibility.js',
     ],
     this.events = const {},
     this.keepAlive = true,
@@ -151,7 +153,6 @@ class HighchartsStock extends StatelessWidget {
       debug: debug,
       keepAlive: keepAlive,
       head: css == null ? null : HighchartsHelpers.styleTag(css),
-      body: javaScriptModules.map(HighchartsHelpers.scriptTag).join('\n'),
       foot: [
         '''
           highcharts_flutter.debug=${debug};
@@ -175,6 +176,11 @@ class HighchartsStock extends StatelessWidget {
               return onLoading!(this);
             }
           : null),
+      javaScriptModules: javaScriptModules,
+    );
+
+    exporting = HighchartsExporting(
+      view: view,
     );
   }
 
@@ -199,6 +205,8 @@ class HighchartsStock extends StatelessWidget {
   /// part to be the actual event name. For example `Highcharts.Chart.load`
   /// represents a listener on `Highcharts.Chart` for the `load` event.
   final Map<String, void Function(Object?)> events;
+
+  late final HighchartsExporting exporting;
 
   final bool keepAlive;